added map icon
